+++ src/arena_tests.c
@@ -5,7 +5,7 @@ void alloc_string_test(void) {
 	set_arena_error_handling_method(ARENA_THROW_ERR);
 	Arena arena = create_arena(32);
 	
-	u64 size = sizeof("hello, world!\n");
+	size_t size = sizeof("hello, world!\n");
 	char *ptr = arena_alloc(&arena, size);
 	arena_copy(&arena, ptr, "hello, world!\n", size);
